mainwindow: Show the availability of the ports and profiles.
If we know if a certain port is available/unavailable, we can print
that out, as a help to the user (and as debugging for ourselves).
A profile is also available/unavailable if all ports which have that
profile are available/unavailable.

Credit goes to David Henningson for the original idea and some of the code.