build-sys: Use autoreconf in bootstrap.sh
Most of the stuff in bootstrap.sh seemed to reimplement autoreconf,
badly. I don't understand the wisdom behind the run_versioned
trickery. configure.ac states that we require Automake 1.11, so
apparently the VERSION variable in bootstrap.sh was out of date.
Instead of bumping the version number, I replaced the cruft with
an autoreconf call.